+++ src/App.jsx
@@ -2,7 +2,10 @@ import { GlobalStyle } from "./shared/styles/GlobalStyle";
 import { Routes } from "react-router-dom";
 import { Route } from "react-router-dom";
 import { SharedLayout } from "./SharedLayout";
-import { lazy } from "react";
+import { lazy, useEffect } from "react";
+import { useDispatch, useSelector } from "react-redux";
+import { selectIsRefreshing } from "./redux/auth/selectors";
+import { refreshUser } from "./redux/auth/auth";
 const SignInPage = lazy(() => import("./pages/SignInPage/SignInPage"));
 const SignUpPage = lazy(() => import("./pages/SignUpPage/SignUpPage"));
 const ForgotPasswordPage = lazy(() =>
@@ -11,16 +14,25 @@ const ForgotPasswordPage = lazy(() =>
 const HomePage = lazy(() => import("./pages/HomePage/HomePage"));
 
 function App() {
-  return (
+  const dispatch = useDispatch();
+  const isRefreshing = useSelector(selectIsRefreshing);
+
+  useEffect(() => {
+    dispatch(refreshUser());
+  }, [dispatch]);
+
+  return isRefreshing ? (
+    <span>Loading...</span>
+  ) : (
     <>
-        <Routes>
-          <Route path="/" element={<SharedLayout />}>
-            <Route index element={<HomePage />} />
-            <Route path="/signup" element={<SignUpPage />} />
-            <Route path="/signin" element={<SignInPage />} />
-            <Route path="/forgot-password" element={<ForgotPasswordPage />} />
-          </Route>
-        </Routes>
+      <Routes>
+        <Route path="/" element={<SharedLayout />}>
+          <Route index element={<HomePage />} />
+          <Route path="/signup" element={<SignUpPage />} />
+          <Route path="/signin" element={<SignInPage />} />
+          <Route path="/forgot-password" element={<ForgotPasswordPage />} />
+        </Route>
+      </Routes>
       <GlobalStyle />
     </>
   );
