+++ src/redux/auth/authSlice.js
@@ -29,7 +29,7 @@ const authSlice = createSlice({
         state.token = null;
         state.isLoggedIn = false;
       })
-      .addCase(auth.refreshUser.pending, state => {
+      .addCase(auth.refreshUser.pending, (state) => {
         state.isRefreshing = true;
       })
       .addCase(auth.refreshUser.fulfilled, (state, action) => {
@@ -37,7 +37,7 @@ const authSlice = createSlice({
         state.isLoggedIn = true;
         state.isRefreshing = false;
       })
-      .addCase(auth.refreshUser.rejected, state => {
+      .addCase(auth.refreshUser.rejected, (state) => {
         state.isRefreshing = false;
       });
   },
