+++ src/redux/user/userOperations.js
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
-import { instance } from "../../shared/api/auth";
+import { instance } from "../auth/auth";
 
 export const getUserInfo = createAsyncThunk(
   "users/getInfo",
